% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{add_edge}
\alias{add_edge}
\alias{add_curved_edge}
\title{Add a single edge between two nodes in a DAG}
\usage{
add_edge(
  dag,
  from,
  to,
  start_position = NULL,
  end_position = NULL,
  options = NULL,
  is_curved = !is.null(curve),
  curve = NULL,
  curve_in_degree = NULL,
  curve_out_degree = NULL,
  is_double_arrow = FALSE,
  is_headless = FALSE,
  annotate = NULL,
  ...
)

add_curved_edge(dag, from, to, options = NULL, curve = "up", ...)
}
\arguments{
\item{dag}{}

\item{from, to}{The names (or numeric id) of nodes create by \code{\link[=add_node]{add_node()}} or related functions.}

\item{start_position, end_position}{The start and end positions of the edge in any manner acceptable to Tikz (e.g., a coordinate in degrees like 30 or a direction like "south west")}

\item{options}{A list of edge options specific to this edge; see \code{\link[=dagtex]{dagtex()}}.}

\item{is_curved}{Logical. Whether the edge should be curved.}

\item{curve}{Direction of the curve, if any: "up" or "down".}

\item{curve_in_degree, curve_out_degree}{Angle in degrees at which the edge leaves and enters the nodes.}

\item{is_double_arrow, is_headless}{Logical. Default is directed arrow; these options can be used to specify headless or two-headed arrows.}

\item{annotate}{Annotation specified with \code{\link[=annotate_edge]{annotate_edge()}}.}

\item{...}{Not currently used.}
}
\value{

}
\description{
Add a single edge between two nodes in a DAG
}
