% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{add_edge}
\alias{add_edge}
\title{Add edges}
\usage{
add_edge(.dag, .from, .to, start_position = NULL, end_position = NULL,
  .options = NULL, is_curved = FALSE, curve = "up",
  is_double_arrow = FALSE, linetype = "solid", annotate = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Add edges
}
\examples{

}
