% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_many_nodes}
\alias{add_many_nodes}
\title{Add several nodes}
\usage{
add_many_nodes(dag, names, options = NULL, ...)
}
\arguments{
\item{dag}{Object of class "dagtex", created with \code{\link[=dagtex]{dagtex()}}.}

\item{names}{Vector of names of the nodes to add. If names within a DAG are
unique, they can also be used when creating edges. If any of the nodes are
split nodes in a SWIG, provide names as a list with nodes to split as
vectors of length 2 (see example).}

\item{options}{List of node options specific to these nodes. For details, see
\code{\link[=dagtex]{dagtex()}}.}

\item{...}{Not in use.}
}
\value{
Object of class "dagtex".
}
\description{
Add several nodes
}
\examples{
dagtex(
  node_options = list(shape = "ellipse"),
  swig_options = list(gap = "3pt", line_color_right = "red")
) \%>\%
  add_many_nodes(list(c("$A_0$", "$a_0$"), "$L_1$", c("$A_1$", "$a_1$"), "$Y$")) \%>\%
  add_many_edges(from = "$a_0$", to = c("$L_1$", "$A_1$", "$Y$")) \%>\%
  add_many_edges(from = "$L_1$", to = c("$A_1$", "$Y$")) \%>\%
  add_edge(from = "$a_1$", to = "$Y$")
}
