% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_node}
\alias{add_node}
\title{Add node(s) to a DAG}
\usage{
add_node(
  dag,
  names,
  options = NULL,
  x = NULL,
  y = NULL,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL,
  adorn_math = NULL,
  ...
)
}
\arguments{
\item{dag}{Object of class "dagtex", created with \code{\link[=dagtex]{dagtex()}}.}

\item{names}{Name of the node, to be printed in DAG.
Can also be a vector of names to add multiple nodes at once, in which case
the first node will be placed at the requested location, and
dagtex() will attempt to place the rest appropriately. For finer control,
add nodes individually. If names within a DAG are
unique, they can also be used when creating edges. If any of the nodes are
split nodes in a SWIG, provide names as a list with nodes to split as
vectors of length 2 (see example).}

\item{options}{List of node options specific to these nodes. For details, see
\code{\link[=dagtex]{dagtex()}}.}

\item{x}{Numeric. Horizontal coordinate at which to place (first) node. If provided,
\code{y} must also be provided.}

\item{y}{Numeric. Vertical coordinate at which to place (first) node. If provided, \code{x}
must also be provided.}

\item{right_of}{Name of node already existing in DAG to place (first) node to
the right of.}

\item{left_of}{Name of node already existing in DAG to place (first) node to
the left of.}

\item{above}{Name of node already existing in DAG to place (first) node
above.}

\item{below}{Name of node already existing in DAG to place (first) node
below.}

\item{adorn_math}{Logical. Node names will be automatically surrounded by
\verb{$...$} if \code{TRUE}. Defaults to \code{FALSE}.}

\item{...}{Not in use.}
}
\value{
Object of class "dagtex".
}
\description{
Add node(s) to a DAG
}
\examples{
dagtex(
  node_options = list(shape = "ellipse"),
  swig_options = list(gap = "3pt", line_color_right = "red")
) \%>\%
  add_node(list(c("$A_0$", "$a_0$"), "$L_1$", c("$A_1$", "$a_1$"), "$Y$")) \%>\%
  add_edge(from = "$a_0$", to = c("$L_1$", "$A_1$", "$Y$")) \%>\%
  add_edge(from = "$L_1$", to = c("$A_1$", "$Y$")) \%>\%
  add_edge(from = "$a_1$", to = "$Y$")
}
