% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_node_to_dag}
\alias{add_node_to_dag}
\title{Add node to the internal DAG structure}
\usage{
add_node_to_dag(
  dag,
  name,
  id,
  coords,
  position,
  options,
  is_swig = FALSE,
  adorn_math = NULL,
  ...
)
}
\description{
Add node to the internal DAG structure
}
\keyword{internal}
