% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R, R/nodes.R
\name{add_curved_edge}
\alias{add_curved_edge}
\alias{annotate_edge}
\alias{add_node}
\alias{add_swig_node}
\title{Add nodes}
\usage{
add_curved_edge(.dag, .from, .to, start_position = NULL,
  end_position = NULL, .options = NULL, curve = "up",
  is_double_arrow = FALSE, linetype = "solid", annotate = NULL, ...)

annotate_edge(text, placement = "midway", position = "above")

add_node(.dag, .name, .options = NULL, x = NULL, y = NULL,
  right_of = NULL, left_of = NULL, above = NULL, below = NULL,
  shape = NULL, is_swig = FALSE, ...)

add_swig_node(.dag, .left, .right, .options = NULL, x = NULL,
  y = NULL, right_of = NULL, left_of = NULL, above = NULL,
  below = NULL, shape = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Add nodes
}
\examples{

}
