% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_swig_node}
\alias{add_swig_node}
\title{Create a split node for a SWIG}
\usage{
add_swig_node(dag, left, right, options = NULL, ...)
}
\arguments{
\item{dag}{Object of class "dagtex", created with \code{\link[=dagtex]{dagtex()}}.}

\item{left}{Name to print on the left (or upper) side of the split node.}

\item{right}{Name to print on the right (or lower) side of the split node.}

\item{options}{List of swig node options specific to this split node. For details, see
\code{\link[=dagtex]{dagtex()}}.}

\item{...}{Other named arguments passed to \code{\link[=add_node]{add_node()}}.}
}
\value{
Object of class "dagtex".
}
\description{
Create a split node for a SWIG
}
\examples{
dagtex() \%>\%
  add_swig_node("A", "a")
# is equivalent to
dagtex() \%>\%
  add_node(c("A", "a"), is_swig = TRUE)
}
