% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{adorn_counterfactuals}
\alias{adorn_counterfactuals}
\title{Add counterfactual notation to SWIG}
\usage{
adorn_counterfactuals(
  dag,
  notation = getOption("dagtex.notation"),
  text_color = NULL
)
}
\arguments{
\item{dag}{Object created by \code{\link[=dagtex]{dagtex()}}.}

\item{notation}{One of "superscript", "subsript", or "parentheses"}

\item{text_color}{Text color}
}
\description{
Add counterfactual notation to SWIG
}
