% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{any_swig_nodes}
\alias{any_swig_nodes}
\title{Check for SWIG nodes}
\usage{
any_swig_nodes(dag)
}
\value{
Logical indicating whether there are any split nodes in the graph.
}
\description{
Check for SWIG nodes
}
\keyword{internal}
