% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{complete}
\alias{complete}
\title{Turn into a complete DAG}
\usage{
complete(.dag, arrow_type = "directed", .options = NULL, ...)
}
\arguments{
\item{arrow_type}{One of "directed", "headless", "double". (Only directed really works with SWIG nodes, the others take some manual manipulation, at least for now.)}
}
\description{
Turn into a complete DAG
}
