% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{complete_dag}
\alias{complete_dag}
\title{Turn DAG into a complete DAG}
\usage{
complete_dag(dag, arrow_type = "directed", options = NULL, ...)
}
\arguments{
\item{arrow_type}{One of "directed", "headless", "double". (Only "directed" really works easily with SWIG nodes, the others take some manual manipulation, at least for now.)}

\item{options}{A list of edge options for each of the new options. Passed to \link{add_edge}.}

\item{...}{Other options passed to \link{add_edge}.}
}
\description{
Turn DAG into a complete DAG
}
