% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{get_latex_code}
\alias{get_latex_code}
\title{Create tikz picture}
\usage{
get_latex_code(dag, add_header = TRUE, node_distance = dag$node_distance)
}
\arguments{
\item{add_header}{}

\item{scale_y}{}
}
\value{



}
\description{
Create tikz picture
}
\examples{
Get LaTeX code


}
