% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_node_coords}
\alias{get_node_coords}
\title{Get coordinates for adding a new node}
\usage{
get_node_coords(
  dag,
  id,
  coords,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL
)
}
\value{
Vector of x,y coordinates at which to place node
}
\description{
Get coordinates for adding a new node
}
\keyword{internal}
