% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{get_node_ids}
\alias{get_node_ids}
\title{Get numeric IDs for nodes}
\usage{
get_node_ids(dag)
}
\arguments{
\item{dag}{Object created by \code{\link[=dagtex]{dagtex()}}.}
}
\value{
A dataframe with columns "name" and "node_id" containing the names
and numeric IDs of the nodes, to be used to reference those nodes
}
\description{
Get numeric IDs for nodes
}
\examples{
dagtex() \%>\%
  add_swig_node("X", "x") \%>\%
  add_node("Y") \%>\%
  get_node_ids()
}
