% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_node_position}
\alias{get_node_position}
\title{Put position-related options into the correct format}
\usage{
get_node_position(
  dag,
  id,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL
)
}
\value{
Tikz code to set location of node (e.g., right = of A)
}
\description{
Put position-related options into the correct format
}
\keyword{internal}
