% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_swigs.R
\name{install_swigs}
\alias{install_swigs}
\title{Install Thomas Richardson's tikz library for drawing SWIGs}
\usage{
install_swigs(...)
}
\description{
Install Thomas Richardson's tikz library for drawing SWIGs
}
\details{
Puts the swigs tikz library into the user's texmf path. This file either
needs to be manually installed, as with this function, or copied from
\code{system.file("tex", "pgflibraryshapesswigs.code.tex", package = "dagtex")}
to the working directory whenever a SWIG is created in an RMarkdown document.
Not necessary if not creating SWIGs, or if only previewing in the viewer and
not via RMarkdown.
}
