% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{knit_printdagtex}
\alias{knit_printdagtex}
\title{Draw DAGs with knitr}
\usage{
knit_printdagtex(
  x,
  density = knitr::opts_current$get("density"),
  fig.path = knitr::opts_current$get("fig.path"),
  savedag = knitr::opts_current$get("savedag"),
  label = knitr::opts_current$get("label"),
  printdag = knitr::opts_current$get("printdag"),
  is_latex_code = knitr::opts_current$get("is_latex_code"),
  has_swig = knitr::opts_current$get("has_swig"),
  ...
)
}
\description{
Draw DAGs with knitr
}
