% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{make_tex_opts}
\alias{make_tex_opts}
\title{Turn options into LaTeX format}
\usage{
make_tex_opts(opts, exclude = c("split"), remove = c("line_type"))
}
\arguments{
\item{exclude}{Options to exclude (character vector)}

\item{remove}{Names (regular expression) of options to remove (i.e., paste
bare argument)}
}
\description{
Turn options into LaTeX format
}
\keyword{internal}
