% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{plot_dagtex}
\alias{plot_dagtex}
\title{Plot DAGs}
\usage{
plot_dagtex(
  dag,
  density = getOption("dagtex.density"),
  is_latex_code = FALSE,
  is_swig = FALSE,
  ...
)
}
\arguments{
\item{dag}{An object created by \code{\link[=dagtex]{dagtex()}} (or the LaTex code generated by
passing such an object to \code{\link[=get_latex_code]{get_latex_code()}}).}

\item{density}{Numeric. Passed to \code{\link[texPreview:tex_preview]{texPreview::tex_preview()}} to control the
density of the image in the Viewer. Large DAGs may need densities smaller
than the default.}

\item{is_latex_code}{Logical. Whether \code{dag} is TikZ code for drawing the DAG
directly. Defaults to (and generally is) \code{FALSE}.}

\item{is_swig}{Logical. Whether the graph is a SWIG (has any split nodes).
Defaults to \code{FALSE}.}

\item{...}{Arguments passed to \code{\link[texPreview:tex_preview]{texPreview::tex_preview()}}.}
}
\value{
A picture in the Viewer.
}
\description{
Plot DAGs
}
\keyword{internal}
