% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{print.dagtex}
\alias{print.dagtex}
\title{Explicitly draw DAG}
\usage{
\method{print}{dagtex}(x, ...)
}
\arguments{
\item{x}{Object created by \code{\link[=dagtex]{dagtex()}}.}

\item{...}{Not currently used.}
}
\description{
Explicitly draw DAG
}
\keyword{internal}
