% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replay_html.magick}
\alias{replay_html.magick}
\title{for building pkgdown site with pics}
\usage{
replay_html.magick(x, fig_save, fig_id, ...)
}
\description{
for building pkgdown site with pics
}
\keyword{internal}
