% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{save_dag}
\alias{save_dag}
\title{Save the DAG image}
\usage{
save_dag(
  dag,
  filename,
  filetype = "png",
  density = getOption("dagtex.density"),
  is_latex_code = FALSE,
  is_swig = FALSE,
  ...
)
}
\arguments{
\item{dag}{An object created by \code{\link[=dagtex]{dagtex()}} (or the LaTex code generated by
passing such an object to \code{\link[=get_latex_code]{get_latex_code()}}).}

\item{filename}{A filepath to create with saved DAG. The filetype (e.g., pdf,
png, jpg) will be extracted from the path if included in the path.}

\item{filetype}{The desired image format, if not provided as part of the
filename. Defaults to 'png', but can be any format accepted by
\code{\link[magick:editing]{magick::image_write()}}.}

\item{density}{Numeric. Passed to \code{\link[texPreview:tex_preview]{texPreview::tex_preview()}} to control the
density of the image.}

\item{is_latex_code}{Logical. Whether \code{dag} is TikZ code for drawing the DAG
directly. Defaults to (and generally is) \code{FALSE}.}

\item{is_swig}{Logical. Whether the graph is a SWIG (has any split nodes).
Defaults to \code{FALSE}.}

\item{...}{Arguments passed to \code{\link[texPreview:tex_preview]{texPreview::tex_preview()}}.}
}
\value{
An image saved to disk; also shows up in the Viewer.
}
\description{
Save the DAG image
}
\examples{
\dontrun{
dagtex() \%>\%
add_many_nodes(c("A", "B", "C")) \%>\%
dag_complete() \%>\%
save_dag("path/to/my/dag.png")
}
}
