% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{split_swig_opts}
\alias{split_swig_opts}
\title{Put SWIG options in the right format}
\usage{
split_swig_opts(swig_options, exclude = "text", ...)
}
\arguments{
\item{exclude}{Text (regular expression) for searching out SWIG options
that use the terms left/right/upper/lower but can't be passed directly}
}
\description{
Put SWIG options in the right format
}
\keyword{internal}
