% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_swigs.R
\name{install_swigs}
\alias{install_swigs}
\alias{use_swigs}
\title{Install Thomas Richardson's tikz library for drawing SWIGs}
\usage{
install_swigs(...)

use_swigs(path = getwd())
}
\arguments{
\item{path}{The directory with the RMarkdown file that uses SWIGs}
}
\description{
Install Thomas Richardson's tikz library for drawing SWIGs

Use Thomas Richardson's tikz library for drawing SWIGs
}
\details{
Puts the swigs tikz library into the user's texmf path. This file either
needs to be manually installed, as with this function, or copied from
\code{system.file("tex", "pgflibraryshapes.swigs.code.tex", package = "dagtex")}
to the working directory whenever a SWIG is created in an RMarkdown document.
Not necessary if not creating SWIGs, or if only previewing in the viewer and
not via RMarkdown.

Copies the swigs tikz library into the specified directory, creating a file
"pgflibraryshapes.swigs.code.tex" to be used when creating SWIGs via
RMarkdown document.
}
