% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{add_edge}
\alias{add_edge}
\alias{add_many_edges}
\alias{add_curved_edge}
\alias{annotate_edge}
\title{Add edges}
\usage{
add_edge(
  .dag,
  .from,
  .to,
  start_position = NULL,
  end_position = NULL,
  .options = NULL,
  curve = NULL,
  is_curved = !is.null(curve),
  curve_in_degree = NULL,
  curve_out_degree = NULL,
  is_double_arrow = FALSE,
  is_headless = FALSE,
  annotate = NULL,
  ...
)

add_many_edges(
  .dag,
  .from,
  .to,
  .options = NULL,
  is_curved = TRUE,
  start_curve = "up",
  ...
)

add_curved_edge(
  .dag,
  .from,
  .to,
  start_position = NULL,
  end_position = NULL,
  .options = NULL,
  curve = "up",
  curve_in_degree = NULL,
  curve_out_degree = NULL,
  is_double_arrow = FALSE,
  is_headless = FALSE,
  annotate = NULL,
  ...
)

annotate_edge(text, placement = "midway", position = "above")
}
\arguments{
\item{.options}{https://gist.github.com/AndiH/f99d9b0cbd3519c27af5b96cfbeff97c}

\item{...}{}
}
\value{

}
\description{
Add edges
}
\examples{

dagtex(.node_options = list(shape = "circle")) \%>\%
 add_node("$A_0$") \%>\%
 add_node("$L_1$") \%>\%
 add_node("$A_1$") \%>\%
 add_node("$Y$") \%>\%
 add_many_edges(.from = "$A_0$", .to = c("$L_1$","$A_1$", "$Y$")) \%>\%
 add_many_edges(.from = "$L_1$", .to = c("$A_1$", "$Y$")) \%>\%
 add_edge(.from = "$A_1$", .to = "$Y$")
}
