% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_node}
\alias{add_node}
\alias{add_many_nodes}
\alias{add_swig_node}
\title{Add nodes}
\usage{
add_node(
  .dag,
  .name,
  .options = NULL,
  x = NULL,
  y = NULL,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL,
  is_swig = FALSE,
  ...
)

add_many_nodes(.dag, .names, .options = NULL, ...)

add_swig_node(
  .dag,
  .left,
  .right,
  .options = NULL,
  x = NULL,
  y = NULL,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add nodes
}
\examples{

dagtex(.node_options = list(shape = "ellipse"),
       .swig_options = list(gap = "3pt", line_color_right = "red")) \%>\%
 add_many_nodes(list(c("$A_0$", "$a_0$"), "$L_1$", c("$A_1$", "$a_1$"), "$Y$")) \%>\%
 add_many_edges(.from = "$a_0$", .to = c("$L_1$","$A_1$", "$Y$")) \%>\%
 add_many_edges(.from = "$L_1$", .to = c("$A_1$", "$Y$")) \%>\%
 add_edge(.from = "$a_1$", .to = "$Y$")
}
