% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_node}
\alias{add_node}
\title{Add a single node to a DAG}
\usage{
add_node(
  dag,
  name,
  options = NULL,
  x = NULL,
  y = NULL,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL,
  is_swig = FALSE,
  adorn_math = NULL,
  ...
)
}
\arguments{
\item{dag}{Object of class "dagtex", created with \code{\link[=dagtex]{dagtex()}}.}

\item{name}{Name of the node, to be printed in DAG. If names within a DAG are
unique, they can also be used when creating edges. If a split node in a
SWIG (\code{is_swig = TRUE}), should be a character vector of length two.}

\item{options}{List of node options specific to this node. For details, see
\code{\link[=dagtex]{dagtex()}}.}

\item{x}{Numeric. Horizontal coordinate at which to place node. If provided,
\code{y} must also be provided.}

\item{y}{Numeric. Vertical coordinate at which to place node.If provided, \code{x}
must also be provided.}

\item{right_of}{Name of node already existing in DAG to place current node to
the right of.}

\item{left_of}{Name of node already existing in DAG to place current node to
the left of.}

\item{above}{Name of node already existing in DAG to place current node
above.}

\item{below}{Name of node already existing in DAG to place current node
below.}

\item{is_swig}{Logical. If node is to be split as part of a SWIG. See also
\code{\link[=add_swig_node]{add_swig_node()}}. Defaults to \code{FALSE}.}

\item{adorn_math}{Logical. Node name will be automatically surrounded by
\verb{$...$} if \code{TRUE}. Defaults to \code{FALSE}.}

\item{...}{Not in use.}
}
\value{

}
\description{
Add a single node to a DAG
}
\examples{

}
