% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{add_node}
\alias{add_node}
\alias{add_swig_node}
\title{Add nodes}
\usage{
add_node(
  .dag,
  .name,
  .options = NULL,
  x = NULL,
  y = NULL,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL,
  is_swig = FALSE,
  ...
)

add_swig_node(
  .dag,
  .left,
  .right,
  .options = NULL,
  x = NULL,
  y = NULL,
  right_of = NULL,
  left_of = NULL,
  above = NULL,
  below = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add nodes
}
\examples{

}
