% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{annotate_edge}
\alias{annotate_edge}
\title{Add text to an edge.}
\usage{
annotate_edge(
  text,
  placement = "midway",
  position = "above",
  size = "normalsize",
  color = NULL
)
}
\arguments{
\item{text}{}

\item{color}{}
}
\description{
Add text to an edge.
}
