% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{annotate_edge}
\alias{annotate_edge}
\title{Add text to an edge.}
\usage{
annotate_edge(
  text,
  placement = "midway",
  position = "above",
  size = "normalsize",
  color = NULL
)
}
\arguments{
\item{text}{Text to add to the DAG.}

\item{placement}{Character, describes where along the edge to place the
annotation. Defaults to "midway", other options include "near start", "at
end", etc. (see TikZ documentation).}

\item{position}{"above", "below", or "" for positioning text on top of the
edge.}

\item{size}{Font size, to LaTex. Defaults to "normalsize", other options are
"footnotesize", "LARGE", etc.}

\item{color}{TExt color of annotation}
}
\description{
Add text to an edge.
}
\examples{
dagtex() \%>\%
add_many_nodes(c("A", "B")) \%>\%
add_edge("A", "B", annotate = annotate_edge("example"))

dagtex() \%>\%
add_many_nodes(c("A", "B")) \%>\%
add_edge("A", "B",
annotate = annotate_edge("example", color = "teal", size = "tiny",
position = "very near start", placement = ""))

}
