% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{get_latex_code}
\alias{get_latex_code}
\title{Get LaTeX code}
\usage{
get_latex_code(dag, add_header = TRUE, ...)
}
\arguments{
\item{dag}{Object created by \code{\link[=dagtex]{dagtex()}}.}

\item{add_header}{Logical. Should the \verb{\\usepackage\{tikz\}} and
\verb{\\usetikzlibrary\{\}} lines be included?}

\item{...}{Not currently used.}
}
\value{
Text that can be used to print the DAG in a LaTeX file.
}
\description{
Generally called internally, can be used to extract the code for the TikZ
drawing to paste directly into a LaTeX document.
}
\examples{
dagtex() \%>\%
  add_node(letters[1:3]) \%>\%
  complete_dag() \%>\%
  get_latex_code()
}
