% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{knit_print.dagtex}
\alias{knit_print.dagtex}
\title{Draw DAGs with knitr}
\usage{
\method{knit_print}{dagtex}(x, options, ...)
}
\arguments{
\item{x}{Object created by \code{\link[=dagtex]{dagtex()}}.}

\item{options}{Additional chunk options include: density (numeric; passed to
\link[texPreview:tex_preview]{texPreview::tex_preview} to control the density of the image for non-pdf
documents); savedag (logical; should the DAG be saved as an additional
file?); printdag (logical; should the DAG be printed?);  is_latex_code
(logical; can be used to pass the code generated by \code{\link[=get_latex_code]{get_latex_code()}}
directly); is_swig (logical; does the TikZ library for SWIGs need to be
called? (Only needed if passing LaTeX code directly.)).}

\item{...}{Not currently used.}
}
\description{
Include DAGs and SWIGs as part of RMarkdown documents by calling \code{\link[=dagtex]{dagtex()}}
in an R code chunk. In pdf documents, the resulting TikZ code will be
included; for all other document types, an image will be printed via
\link[texPreview:tex_preview]{texPreview::tex_preview}. This function is not called directly.
}
\keyword{internal}
