% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{knit_print.dagtex}
\alias{knit_print.dagtex}
\title{Draw DAGs with knitr}
\usage{
\method{knit_print}{dagtex}(
  x,
  density = knitr::opts_current$get("density"),
  fig.path = knitr::opts_current$get("fig.path"),
  ...
)
}
\description{
Draw DAGs with knitr
}
