% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_swigs.R
\name{use_swigs}
\alias{use_swigs}
\title{Use Thomas Richardson's tikz library for drawing SWIGs}
\usage{
use_swigs(path = getwd())
}
\arguments{
\item{path}{The directory with the RMarkdown file that uses SWIGs}
}
\description{
Use Thomas Richardson's tikz library for drawing SWIGs
}
\details{
Copies the swigs tikz library into the specified directory, creating a file
"pgflibraryshapes.swigs.code.tex" to be used when creating SWIGs via
RMarkdown document.
}
